Docstrings vs multiline string literals distinction for pyblocks
Pyblocks will now detect if a block as a whole is just a single
multiline string literal. If it is, it will be reindented in the
same way as a usual block and stripped of its trailing spaces if
that was requested. If it isn't, but it contains multiline
string definitions somewhere inside, the definitions will not be
reindented or stripped of trailing spaces. That way, the semantics
of the code is guaranteed to be preserved.

	modified:   pyblocks.py